Began creating tests for User Service class to guide creation of building methods.  Completed deleteUserTest 1

+++ BankingApp/src/main/java/com/banking/BankingApp/repository/UserRepository.java
@@ -2,6 +2,9 @@ package com.banking.BankingApp.repository;
 import com.banking.BankingApp.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
+
+
+
 
 }
